Added multi-object rendering.
Doing this required rewriting a lot of other code.

Packed instance data into the object struct. (Necessary for multi-object rendering.)

Temporarily deprecated model rendering code; slight chance of action permanence. (The code relies on external instancing set-up, which is a problem.)

The two Junks are preparing to form a staircase.