Fixed Vsync.
I was rendering 900+ frames per second on my machine. It might have been more, but I didn't wait long enough to get an accurate reading.

Deleted the '2' file. It serves no purpose and was created by accident, but I feel a little bad for deleting it.

The Junks have become a star system.