Implemented preliminary UI rendering.
Atrocious (and straight out wrong) code that will react to any change by just not working. In addition, the UI element is backwards; doing absolutely anything to try and fix that will result in nothing being rendered, so a bunch of strange values that seemingly make no sense are used.

The Junks now have something to stare at, although I'm not sure if they actually see it as their faces causes the background element to disappear...